+++ src/App.css
@@ -31,12 +31,6 @@ a {
   text-decoration-thickness: 75px;
 }
 
-.logo {
-  height: 6em;
-  padding: 1.5em;
-  will-change: filter;
-  transition: filter 300ms;
-}
 .logo:hover {
   filter: drop-shadow(0 0 2em #646cffaa);
 }
