+++ src/index.css
@@ -1,5 +1,14 @@
 :root {
-  font-family: system-ui, Avenir, Helvetica, Arial, sans-serif;
+  width: 100%;
+  height: 100%;
+  align-items: center;
+  justify-content: center;
+  box-sizing: border-box;
+  margin: 0;
+  padding: 0;
+  font-size: 16px;
+
+  font-family: system-ui, Helvetica, Arial, sans-serif, sys;
   line-height: 1.5;
   font-weight: 400;
 
@@ -20,6 +29,12 @@ a {
 }
 a:hover {
   color: #535bf2;
+}
+
+.goToMap {
+  text-align: center;
+  margin-top: 20px;
+  margin-bottom: 20px;
 }
 
 body {
