+++ src/main.jsx
@@ -5,8 +5,7 @@ import { BrowserRouter, Routes, Route } from "react-router";
 
 import "./index.css";
 import App from "./App.jsx";
-import HealthcareAccess from "./HealthcareAccess";
-import AirQuality from "./AirQuality";
+import DeckComponent from "./DeckComponent.jsx";
 
 const root = document.getElementById("root");
 
@@ -15,8 +14,7 @@ createRoot(document.getElementById("root")).render(
     <BrowserRouter>
       <Routes>
         <Route path="/" element={<App />} />
-        <Route path="healthcareaccess" element={<HealthcareAccess />} />
-        <Route path="airquality" element={<AirQuality />} />
+        <Route path="deckcomponent" element={<DeckComponent />} />
       </Routes>
     </BrowserRouter>
   </StrictMode>
